Create base class for form iterating
