+++ corehq/apps/app_manager/management/commands/check_for_form_entities.py
@@ -1,77 +1,20 @@
-import re
-import signal
 from datetime import datetime
 from couchdbkit.exceptions import ResourceNotFound
 from django.conf import settings
-from django.core.management.base import BaseCommand
-from django.core.paginator import Paginator
 from django.urls import reverse
-from jsonobject.exceptions import BadValueError, WrappingAttributeError
-from corehq.apps.app_manager.models import Application, GlobalAppConfig
 from corehq.apps.app_manager.xform import parse_xml
 from corehq.apps.app_manager.exceptions import DangerousXmlException, XFormException
-from dimagi.utils.couch.database import iter_docs
-from corehq.util.log import with_progress_bar
-from corehq.util.doc_processor.couch import resumable_view_iterator
-from lxml.html.clean import Cleaner
+
+from .form_iterator import FormIteratorCommandBase, get_current_apps, get_forms
 
 
-class Command(BaseCommand):
-    help = '''
-    Searches through application forms to determine if XML entities exist.
-    Resumable. Call with --reset if you wish to start a fresh run
-    '''
+class Command(FormIteratorCommandBase):
 
     def __init__(self):
         self.entities_found = 0
         self.total_forms = 0
         self.start_time = None
-        self.xss_attempt_found = 0
-
-    def add_arguments(self, parser):
-        parser.add_argument('--all', action='store_true',
-            help='check all applications, rather than only current ones')
-        parser.add_argument('--limit', default=0, type=int, help='terminate after LIMIT forms')
-        parser.add_argument('--reset', action='store_true', help='trigger a fresh run')
-        parser.add_argument('--path', default='output.log', help='the path to write log output')
-        parser.add_argument('--batchsize', default=10, type=int, help='number of apps to pull simultaneously')
-        #this is temp
-        parser.add_argument('--xss', action='store_true', help="check for xss instead of xxe")
-
-    def handle(self, all, limit, xss, reset, path, batchsize, *args, **kwargs):
-        write_mode = 'w' if reset else 'a'  # Only overwrite on fresh runs
-
-        # NOTE: Outputting to a file, rather than stdout for some messages due to the progress bar.
-        # Because the progress bar alters stdout, just using 'tee' doesn't produce a good log file,
-        # Hence the support for a separate log file here
-        with open(path, write_mode) as f:
-            self.log_file = f
-            self.setup_interrupt_handler()
-
-            if all:
-                self.broadcast('Searching for entities in ALL forms')
-            else:
-                self.broadcast('Searching for entities in CURRENT forms')
-
-            self.start_time = datetime.now()
-            if xss:
-                self.find_XSS(all, limit, reset, batchsize)
-                self.report_results(True)
-            else:
-                self.process_forms(all, limit, reset, batchsize)
-                self.report_results()
-
-    def find_XSS(self, all, limit, reset, batchsize):
-        get_apps = self.get_all_apps if all else get_current_apps
-
-        for app in get_apps(reset, batchsize):
-            for form in get_forms(app):
-                if form_contains_xss_attempt(form):
-                    handle_entity_form(form, app, self.log_file, True)
-                    self.xss_attempt_found += 1
-                self.total_forms += 1
-                if limit > 0 and self.total_forms >= limit:
-                    return
+        self.iteration_key = __name__
 
     def process_forms(self, all, limit, reset, batchsize):
         get_apps = self.get_all_apps if all else get_current_apps
@@ -85,116 +28,10 @@ class Command(BaseCommand):
                 if limit > 0 and self.total_forms >= limit:
                     return
 
-    def report_results(self, xss=False):
+    def report_results(self):
         elapsed = datetime.now() - self.start_time
-        if xss:
-            self.broadcast(f'Found {self.xss_attempt_found} XSS attempts in {self.total_forms} forms')
-        else:
-            self.broadcast(f'Found {self.entities_found} entities in {self.total_forms} forms')
+        self.broadcast(f'Found {self.entities_found} entities in {self.total_forms} forms')
         self.broadcast(f'Completed search in {elapsed}')
-
-    def setup_interrupt_handler(self):
-        def interrupt_handler(signum, frame):
-            self.report_results()
-            exit(1)
-
-        signal.signal(signal.SIGINT, interrupt_handler)
-
-    def broadcast(self, msg):
-        '''Send a message to both stdout and the file'''
-        print(msg)
-        print(msg, file=self.log_file)
-
-    def get_all_apps(self, reset=False, batchsize=10):
-        '''
-        This looks at all apps, including previous versions.
-        Note that this won't look at linked or remote apps
-        '''
-        db = Application.get_db()
-        keys = [[Application.__name__], [f'{Application.__name__}-Deleted']]
-        view_name = 'all_docs/by_doc_type'
-        raw_iter = resumable_view_iterator(db, __name__, view_name, view_keys=keys, chunk_size=batchsize,
-            full_row=True)
-        if reset:
-            raw_iter.discard_state()
-
-        modified_start_key, keys = get_keys_to_search(keys, raw_iter)
-
-        count = get_remaining_app_count(db, view_name, keys, modified_start_key)
-
-        app_iter = (wrapped_app for wrapped_app in (self.wrap_app(x) for x in raw_iter) if wrapped_app)
-        return with_progress_bar(app_iter, count)
-
-    def wrap_app(self, fetched_row):
-        id = fetched_row['id']
-        raw_app = fetched_row['doc']
-        if not raw_app:
-            print(f'no associated app found for: {id}', file=self.log_file)
-            return None
-
-        try:
-            return Application.wrap(raw_app)
-        except (BadValueError, WrappingAttributeError, ValueError, AssertionError):
-            print(f'could not wrap app: {raw_app["_id"]}', file=self.log_file)
-            return None
-
-
-def get_keys_to_search(keys, iterator):
-    # Check for existing progress with the resumable iterator
-    # This is working with the assumption that the iterator will iterate through the keys in the order
-    # they are specified. This is brittle, but I don't have a better solution without reworking
-    # the resumable iterator. This method knows FAR too much of the internals of the resumable iterator
-    if 'startkey' not in iterator.state.kwargs:
-        return (None, keys)
-
-    startkey = list(iterator.state.kwargs['startkey'])
-    base_start = startkey[:-1]
-    index = keys.index(base_start)
-    keys = keys[index:]
-    return (startkey, keys)
-
-
-def get_remaining_app_count(db, view_name, keys, modified_start_key):
-    count = 0
-    for index, key in enumerate(keys):
-        if index == 0 and modified_start_key:
-            startkey = modified_start_key
-        else:
-            startkey = key
-
-        endkey = key + [{}]
-        query = db.view(view_name, reduce=True, startkey=startkey, endkey=endkey)
-        count += query.first()['value']
-
-    return count
-
-
-def get_current_apps(reset=False, batchsize=10):
-    '''
-    Only examine the most recent version of any application. This does look at linked and remote apps.
-    Note that this doesn't support resumable execution, as the effort was out of scope,
-    and therefore reset is unused.
-    '''
-    query = GlobalAppConfig.objects.values_list('app_id', flat=True).order_by('id')
-    count = query.count()
-    iter = get_current_apps_iter(query, batchsize)
-
-    return with_progress_bar(iter, count)
-
-
-def get_current_apps_iter(query, batchsize):
-    db = Application.get_db()
-    paginator = Paginator(query, 100)
-    for page_num in paginator.page_range:
-        page = paginator.page(page_num)
-        app_ids = list(page.object_list)
-        for app_doc in iter_docs(db, app_ids, chunksize=batchsize):
-            yield Application.wrap(app_doc)
-
-
-def get_forms(app):
-    for module in app.get_modules():
-        yield from module.get_forms()
 
 
 def form_contains_entities(form):
@@ -213,11 +50,8 @@ def form_contains_entities(form):
     return False
 
 
-def handle_entity_form(form, app, log_file, xss=False):
-    if xss:
-        print('Found XSS attempt', file=log_file)
-    else:
-        print('Found entity', file=log_file)
+def handle_entity_form(form, app, log_file):
+    print('Found entity', file=log_file)
     print(f'\tForm: {form.unique_id}, app {app._id}', file=log_file)
     if app.copy_of:
         print(f'\tApp is a copy of {app.copy_of}. Found on version {app.version}', file=log_file)
@@ -225,42 +59,3 @@ def handle_entity_form(form, app, log_file, xss=False):
     host = settings.BASE_ADDRESS
     url = reverse('get_xform_source', args=(app.domain, app._id, form.unique_id,))
     print(f'\tview document at: {protocol}://{host}{url}', file=log_file)
-
-
-def form_contains_xss_attempt(form):
-    """
-    This is a test run - it'll likely change!!
-
-    The idea is to have lxml clean an html snippet (user input) with and without
-    the function that eliminates javascript/unsafe attributes and compare the two to
-    identify any html that did have those things.
-    """
-
-    regex = '(?<=<value>)(.*?)(?=\s*<\/value>)'
-    regexoutput = '(?<=<output)(.*?)(?=\s*>)'
-
-    cleaner = Cleaner()
-    clean_html = cleaner.clean_html
-    cleaner_js = Cleaner(javascript=False, safe_attrs_only=False)
-    cleaner_html_js = cleaner_js.clean_html
-    #this parts copied
-    try:
-        source = form.source
-    except ResourceNotFound:
-        return False
-
-    for value in re.findall(regex, form.source):
-        if '&lt;' and '&gt;' in value:
-            value = value.replace("&lt;", "<").replace("&gt;", ">")
-        if '<output' in value:
-            for output in re.findall(regexoutput, value):
-                value = value.replace(output, '')
-            value = value.replace("<output>", '').replace("</output>", '')
-
-        cleaned_js = cleaner_html_js("<div>" + value + "</div>")[5:-6]
-        cleaned = clean_html("<div>" + value + "</div>")[5:-6]
-
-        if cleaned != cleaned_js:
-            print(cleaned_js)
-            return True
-    return False
